Rework tests to be based on pyatspi only
dogtail is not packaged for ubuntu at the moment and dogtail adds
another layer which can make it difficult to debug intermittent
failures. We might switch back at some point but for now I think
it's easier to stay lower level and try to figure out how to
improve reliability there.